VACMS-13317: Fix saving taxonomy terms, other things.
