+++ docroot/modules/custom/va_gov_backend/src/Plugin/Validation/Constraint/PreventAbsoluteCmsLinksValidator.php
@@ -3,6 +3,7 @@
 namespace Drupal\va_gov_backend\Plugin\Validation\Constraint;
 
 use Drupal\Component\Utility\Html;
+use Drupal\va_gov_backend\Plugin\Validation\Constraint\Traits\TextValidatorTrait;
 use Drupal\va_gov_backend\Plugin\Validation\Constraint\Traits\ValidatorContextAccessTrait;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
@@ -12,53 +13,13 @@ use Symfony\Component\Validator\ConstraintValidator;
  */
 class PreventAbsoluteCmsLinksValidator extends ConstraintValidator {
 
+  use TextValidatorTrait;
   use ValidatorContextAccessTrait;
-
-  /**
-   * Add a violation message.
-   *
-   * @param int $delta
-   *   The field delta.
-   * @param string $message
-   *   The violation message.
-   * @param array $params
-   *   Message parameters.
-   */
-  public function addViolation(int $delta, string $message, array $params = []) {
-    $this->getContext()
-      ->buildViolation($message, $params)
-      ->atPath((string) $delta . '.value')
-      ->addViolation();
-  }
 
   /**
    * {@inheritdoc}
    */
-  public function validate($items, Constraint $constraint) {
-    foreach ($items as $delta => $item) {
-      $type = $item->getFieldDefinition()->getType();
-      $fieldValue = $item->getValue();
-      /** @var \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventAbsoluteCmsLinks $constraint */
-      if ($type === 'text_long' && $fieldValue['format'] !== 'plain_text') {
-        $this->validateHtml($fieldValue['value'], $constraint, $delta);
-      }
-      else {
-        $this->validateText($fieldValue['value'], $constraint, $delta);
-      }
-    }
-  }
-
-  /**
-   * Validates plain text.
-   *
-   * @param string $text
-   *   A plain text string to validate.
-   * @param \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventAbsoluteCmsLinks $constraint
-   *   The constraint we're validating.
-   * @param int $delta
-   *   The field item delta.
-   */
-  public function validateText(string $text, PreventAbsoluteCmsLinks $constraint, int $delta) {
+  public function validateText(string $text, Constraint $constraint, int $delta) {
     if (strpos($text, 'cms.va.gov') === FALSE) {
       return;
     }
@@ -71,16 +32,10 @@ class PreventAbsoluteCmsLinksValidator extends ConstraintValidator {
   }
 
   /**
-   * Validates HTML.
-   *
-   * @param string $html
-   *   An HTML string to validate.
-   * @param \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventAbsoluteCmsLinks $constraint
-   *   The constraint we're validating.
-   * @param int $delta
-   *   The field item delta.
+   * {@inheritdoc}
    */
-  public function validateHtml(string $html, PreventAbsoluteCmsLinks $constraint, int $delta) {
+  public function validateHtml(string $html, Constraint $constraint, int $delta) {
+    /** @var \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventAbsoluteCmsLinks $constraint */
     if (strpos($html, 'cms.va.gov') === FALSE) {
       return;
     }
