+++ docroot/modules/custom/va_gov_backend/src/Plugin/Validation/Constraint/PreventAbsoluteUrlsAsPathsInLinksValidator.php
@@ -3,6 +3,7 @@
 namespace Drupal\va_gov_backend\Plugin\Validation\Constraint;
 
 use Drupal\Component\Utility\Html;
+use Drupal\va_gov_backend\Plugin\Validation\Constraint\Traits\TextValidatorTrait;
 use Drupal\va_gov_backend\Plugin\Validation\Constraint\Traits\ValidatorContextAccessTrait;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
@@ -12,53 +13,13 @@ use Symfony\Component\Validator\ConstraintValidator;
  */
 class PreventAbsoluteUrlsAsPathsInLinksValidator extends ConstraintValidator {
 
+  use TextValidatorTrait;
   use ValidatorContextAccessTrait;
-
-  /**
-   * Add a violation message.
-   *
-   * @param int $delta
-   *   The field delta.
-   * @param string $message
-   *   The violation message.
-   * @param array $params
-   *   Message parameters.
-   */
-  public function addViolation(int $delta, string $message, array $params = []) {
-    $this->getContext()
-      ->buildViolation($message, $params)
-      ->atPath((string) $delta . '.value')
-      ->addViolation();
-  }
 
   /**
    * {@inheritdoc}
    */
-  public function validate($items, Constraint $constraint) {
-    foreach ($items as $delta => $item) {
-      $type = $item->getFieldDefinition()->getType();
-      $fieldValue = $item->getValue();
-      /** @var \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventAbsoluteUrlsAsPathsInLinks $constraint */
-      if ($type === 'text_long' && $fieldValue['format'] !== 'plain_text') {
-        $this->validateHtml($fieldValue['value'], $constraint, $delta);
-      }
-      else {
-        $this->validateText($fieldValue['value'], $constraint, $delta);
-      }
-    }
-  }
-
-  /**
-   * Validates plain text.
-   *
-   * @param string $text
-   *   A plain text string to validate.
-   * @param \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventAbsoluteUrlsAsPathsInLinks $constraint
-   *   The constraint we're validating.
-   * @param int $delta
-   *   The field item delta.
-   */
-  public function validateText(string $text, PreventAbsoluteUrlsAsPathsInLinks $constraint, int $delta) {
+  public function validateText(string $text, Constraint $constraint, int $delta) {
     /*
      * Regular expression explanation:
      *
@@ -76,6 +37,7 @@ class PreventAbsoluteUrlsAsPathsInLinksValidator extends ConstraintValidator {
      * In other words, we look for a string that looks like a valid URL but is
      * immediately preceded by a forward slash.
      */
+    /** @var \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventAbsoluteUrlsAsPathsInLinks $constraint */
     if (preg_match('#(/https?://[^\s]+)#', $text, $matches)) {
       $this->addViolation($delta, $constraint->plainTextMessage, [
         ':url' => $matches[1],
@@ -84,16 +46,10 @@ class PreventAbsoluteUrlsAsPathsInLinksValidator extends ConstraintValidator {
   }
 
   /**
-   * Validates HTML.
-   *
-   * @param string $html
-   *   An HTML string to validate.
-   * @param \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventAbsoluteUrlsAsPathsInLinks $constraint
-   *   The constraint we're validating.
-   * @param int $delta
-   *   The field item delta.
+   * {@inheritdoc}
    */
-  public function validateHtml(string $html, PreventAbsoluteUrlsAsPathsInLinks $constraint, int $delta) {
+  public function validateHtml(string $html, Constraint $constraint, int $delta) {
+    /** @var \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventAbsoluteUrlsAsPathsInLinks $constraint */
     $dom = Html::load($html);
     $xpath = new \DOMXPath($dom);
     foreach ($xpath->query('//a[starts-with(@href, "/http")]') as $element) {
