+++ docroot/modules/custom/va_gov_backend/src/Plugin/Validation/Constraint/PreventProtocolRelativeLinksValidator.php
@@ -3,6 +3,7 @@
 namespace Drupal\va_gov_backend\Plugin\Validation\Constraint;
 
 use Drupal\Component\Utility\Html;
+use Drupal\va_gov_backend\Plugin\Validation\Constraint\Traits\TextValidatorTrait;
 use Drupal\va_gov_backend\Plugin\Validation\Constraint\Traits\ValidatorContextAccessTrait;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
@@ -12,53 +13,13 @@ use Symfony\Component\Validator\ConstraintValidator;
  */
 class PreventProtocolRelativeLinksValidator extends ConstraintValidator {
 
+  use TextValidatorTrait;
   use ValidatorContextAccessTrait;
-
-  /**
-   * Add a violation message.
-   *
-   * @param int $delta
-   *   The field delta.
-   * @param string $message
-   *   The violation message.
-   * @param array $params
-   *   Message parameters.
-   */
-  public function addViolation(int $delta, string $message, array $params = []) {
-    $this->getContext()
-      ->buildViolation($message, $params)
-      ->atPath((string) $delta . '.value')
-      ->addViolation();
-  }
 
   /**
    * {@inheritdoc}
    */
-  public function validate($items, Constraint $constraint) {
-    foreach ($items as $delta => $item) {
-      $type = $item->getFieldDefinition()->getType();
-      $fieldValue = $item->getValue();
-      /** @var \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventProtocolRelativeLinks $constraint */
-      if ($type === 'text_long' && $fieldValue['format'] !== 'plain_text') {
-        $this->validateHtml($fieldValue['value'], $constraint, $delta);
-      }
-      else {
-        $this->validateText($fieldValue['value'], $constraint, $delta);
-      }
-    }
-  }
-
-  /**
-   * Validates plain text.
-   *
-   * @param string $text
-   *   A plain text string to validate.
-   * @param \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventProtocolRelativeLinks $constraint
-   *   The constraint we're validating.
-   * @param int $delta
-   *   The field item delta.
-   */
-  public function validateText(string $text, PreventProtocolRelativeLinks $constraint, int $delta) {
+  public function validateText(string $text, Constraint $constraint, int $delta) {
     /*
      * Regular expression explanation:
      *
@@ -96,6 +57,7 @@ class PreventProtocolRelativeLinksValidator extends ConstraintValidator {
      * - contains at least two "words" separated by dots
      * - may contain arbitrary non-whitespace characters after that.
      */
+    /** @var \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventProtocolRelativeLinks $constraint */
     if (preg_match('#([^(\w+:)]//[^/]+\w+(\.\w+)+\S*)#', $text, $matches)) {
       $this->addViolation($delta, $constraint->plainTextMessage, [
         ':url' => $matches[1],
@@ -104,16 +66,10 @@ class PreventProtocolRelativeLinksValidator extends ConstraintValidator {
   }
 
   /**
-   * Validates HTML.
-   *
-   * @param string $html
-   *   An HTML string to validate.
-   * @param \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventProtocolRelativeLinks $constraint
-   *   The constraint we're validating.
-   * @param int $delta
-   *   The field item delta.
+   * {@inheritdoc}
    */
-  public function validateHtml(string $html, PreventProtocolRelativeLinks $constraint, int $delta) {
+  public function validateHtml(string $html, Constraint $constraint, int $delta) {
+    /** @var \Drupal\va_gov_backend\Plugin\Validation\Constraint\PreventProtocolRelativeLinks $constraint */
     $dom = Html::load($html);
     $xpath = new \DOMXPath($dom);
     foreach ($xpath->query('//a[starts-with(@href, "//")]') as $element) {
